Fixed Sonar bug

+++ src/com/simplicite/commons/Demo/DemoCommon.java
@@ -50,7 +50,7 @@ public class DemoCommon implements java.io.Serializable {
 
 		// Stock is considered low if less than X% of total ordered quantity
 		AppLog.info(DemoCommon.class, "isLowStock", "Current stock = " + stock, grant);
-		int threshold = Math.round((grant.getIntParameter("DEMO_PRD_LOWSTOCK_THRESHOLD", 10) / 100) * quantity);
+		int threshold = Math.round(((float)grant.getIntParameter("DEMO_PRD_LOWSTOCK_THRESHOLD", 10) / 100) * quantity);
 		AppLog.info(DemoCommon.class, "isLowStock", "Low stock threshold " + threshold, grant);
 		boolean lowStock = stock < threshold;
 		AppLog.info(DemoCommon.class, "isLowStock", "Low stock " + lowStock, grant);
